refactor(composer_rpc): move combined delivery simulation to delivery.rs
Move simulate_l1_combined_delivery (542 LOC),
simulate_chained_delivery_l2_to_l1 (225 LOC), and
fallback_per_call_l2_to_l1_simulation (64 LOC) from l2_to_l1.rs
to the shared delivery.rs module.

Cross-module call to enrich_return_calls_via_l2_trace (stays in
l2_to_l1.rs) via super::l2_to_l1::enrich_return_calls_via_l2_trace.

l2_to_l1.rs: 4170 → 3301 (-869 lines).
delivery.rs: 1132 → 1995 (+863 lines).

Progress: l1_to_l2 4345→3828, l2_to_l1 5459→3301.

616/616 tests, clippy clean.
