+++ crates/based-rollup/src/composer_rpc/delivery.rs
@@ -1,9 +1,9 @@
 //! L1 delivery simulation for L2→L1 cross-chain calls.
 //!
-//! Contains the core `simulate_l1_delivery` function and its helper functions,
-//! shared between `l2_to_l1.rs` (single-call path) and `direction.rs`
-//! (L2ToL1::enrich_calls_before_retrace). Also provides helpers used by
-//! `simulate_l1_combined_delivery` in `l2_to_l1.rs`.
+//! Contains the core `simulate_l1_delivery` function (single-call),
+//! `simulate_l1_combined_delivery` (multi-call), and their helper functions.
+//! Shared between `l2_to_l1.rs` and `direction.rs`
+//! (L2ToL1::enrich_calls_before_retrace).
 
 use alloy_primitives::{Address, U256};
 use alloy_sol_types::SolCall;
@@ -1129,4 +1129,867 @@ pub(crate) async fn build_and_run_l1_postbatch_trace(
 
     let user_trace = bundle_traces[1].clone();
     Some((user_trace, resp))
+}
+
+/// Simulate ALL L2→L1 calls in one `debug_traceCallMany` bundle so later calls can see
+/// state effects from earlier ones (e.g., tokens minted by call_0 available to call_1).
+///
+/// This is the multi-call counterpart of `simulate_l1_delivery`. Instead of simulating
+/// each call independently, it builds a combined bundle:
+///   `[postBatch(combined_entries), createProxy(user0), trigger0, createProxy(user1), trigger1, ...]`
+///
+/// Uses the same iterative discovery loop: if trigger traces reveal new return calls,
+/// entries are rebuilt with continuation structure and re-simulated until convergence
+/// or `MAX_SIMULATION_ITERATIONS`.
+///
+/// Returns `None` if the simulation cannot be performed. Otherwise returns a vec of
+/// per-call results: `(delivery_return_data, delivery_failed, detected_return_calls)`.
+#[allow(clippy::too_many_arguments)]
+pub(crate) async fn simulate_l1_combined_delivery(
+    client: &reqwest::Client,
+    l1_rpc_url: &str,
+    l2_rpc_url: &str,
+    cross_chain_manager_address: Address,
+    rollups_address: Address,
+    builder_address: Address,
+    builder_private_key: Option<&str>,
+    rollup_id: u64,
+    calls: &[&DiscoveredCall],
+    rlp_encoded_tx: &[u8],
+) -> Option<Vec<(Vec<u8>, bool, Vec<ReturnEdge>)>> {
+    if calls.is_empty() {
+        return Some(vec![]);
+    }
+
+    // Check all destinations have code on L1.
+    for call in calls {
+        let code_req = serde_json::json!({
+            "jsonrpc": "2.0",
+            "method": "eth_getCode",
+            "params": [format!("{}", call.destination), "latest"],
+            "id": 1
+        });
+        let code_resp = client.post(l1_rpc_url).json(&code_req).send().await.ok()?;
+        let code_body: Value = code_resp.json().await.ok()?;
+        let code_hex = code_body.get("result")?.as_str()?;
+        if code_hex == "0x" || code_hex == "0x0" {
+            tracing::info!(
+                target: "based_rollup::proxy",
+                destination = %call.destination,
+                "L1 target is EOA — skipping combined simulation"
+            );
+            return None;
+        }
+    }
+
+    tracing::info!(
+        target: "based_rollup::proxy",
+        num_calls = calls.len(),
+        "running combined L1 delivery simulation for multi-call L2→L1"
+    );
+
+    // Parse builder private key — required for signing postBatch proof.
+    let key_hex = builder_private_key?;
+    let key_clean = key_hex.strip_prefix("0x").unwrap_or(key_hex);
+    let builder_key = match key_clean.parse::<alloy_signer_local::PrivateKeySigner>() {
+        Ok(k) => k,
+        Err(e) => {
+            tracing::warn!(
+                target: "based_rollup::proxy",
+                %e,
+                "failed to parse builder private key for combined L1 simulation"
+            );
+            return None;
+        }
+    };
+
+    // Iterative discovery loop.
+    let mut all_return_calls: Vec<ReturnEdge> = Vec::new();
+    // Per-call results: seed from known delivery data instead of empty placeholders.
+    // The L2 iterative discovery already captured real delivery return data via its
+    // own L1 simulation. Using it here gives correct RESULT hashes on iteration 1,
+    // eliminating the placeholder→real→confirm convergence cycle for leaf deliveries.
+    let mut per_call_return_data: Vec<Vec<u8>> = calls
+        .iter()
+        .map(|c| c.delivery_return_data.clone())
+        .collect();
+    let mut per_call_delivery_failed: Vec<bool> = calls.iter().map(|c| c.delivery_failed).collect();
+    // Track previous return data for convergence (#254 item 7).
+    let mut prev_per_call_return_data: Vec<Vec<u8>> = per_call_return_data.clone();
+
+    for iteration in 1..=MAX_SIMULATION_ITERATIONS {
+        tracing::info!(
+            target: "based_rollup::proxy",
+            iteration,
+            known_return_calls = all_return_calls.len(),
+            "combined L1 delivery simulation iteration"
+        );
+
+        // Build combined L1 deferred entries for all calls.
+        // For multi-call patterns (siblings), each call needs scope=[..., sibling_index]
+        // for executeL2TX to route between them via scope navigation.
+        let is_multi_call = calls.len() > 1;
+        let mut combined_entries: Vec<cross_chain::CrossChainExecutionEntry> = Vec::new();
+        for (i, call) in calls.iter().enumerate() {
+            // Collect return calls belonging to this trigger call.
+            let my_return_calls: Vec<&ReturnEdge> = all_return_calls
+                .iter()
+                .filter(|rc| rc.parent_call_index == cross_chain::ParentLink::Child(cross_chain::AbsoluteCallIndex::new(i)))
+                .collect();
+
+            // Scope: for single calls, use trace_depth-1 (direct caller excluded).
+            // For multi-call (siblings), append sibling_index for routing.
+            let call_scope = if is_multi_call {
+                let mut s = if call.trace_depth <= 1 {
+                    vec![]
+                } else {
+                    vec![U256::ZERO; call.trace_depth]
+                };
+                s.push(U256::from(i));
+                s
+            } else if call.trace_depth <= 1 {
+                vec![]
+            } else {
+                vec![U256::ZERO; call.trace_depth]
+            };
+
+            let entries = if my_return_calls.is_empty() {
+                // Simple case: just this L2→L1 call.
+                let call_entries = cross_chain::build_l2_to_l1_call_entries(
+                    call.destination,
+                    call.calldata.to_vec(),
+                    call.value,
+                    call.source_address,
+                    rollup_id,
+                    rlp_encoded_tx.to_vec(),
+                    per_call_return_data[i].clone(),
+                    per_call_delivery_failed[i],
+                    call_scope.clone(),
+                    cross_chain::TxOutcome::Success, // tx_reverts
+                );
+                call_entries.l1_deferred_entries
+            } else {
+                // Continuation case: use the SAME table builder functions as the
+                // real batch to ensure identical entry ordering (same fix as
+                // simulate_l1_delivery).
+                let root_call = crate::table_builder::L2DetectedCall {
+                    destination: call.destination,
+                    data: call.calldata.to_vec(),
+                    value: call.value,
+                    source_address: call.source_address,
+                    delivery_return_data: per_call_return_data[i].clone(),
+                    delivery_failed: per_call_delivery_failed[i],
+                    scope: ScopePath::from_parts(call_scope.clone()),
+                    in_reverted_frame: false,
+                };
+
+                let return_calls_for_builder: Vec<crate::table_builder::L2ReturnCall> =
+                    my_return_calls
+                        .iter()
+                        .map(|rc| crate::table_builder::L2ReturnCall {
+                            destination: rc.destination,
+                            data: rc.data.clone(),
+                            value: rc.value,
+                            source_address: rc.source_address,
+                            // parent_call_index in ReturnEdge refers to the
+                            // combined simulation's call index. For the table builder,
+                            // we're building entries for a single root call, so set to
+                            // None (defaults to last L2→L1 call, which is the only one).
+                            parent_call_index: cross_chain::ParentLink::Root,
+                            l2_return_data: rc.return_data.clone(),
+                            l2_delivery_failed: rc.delivery_failed,
+                            scope: rc.scope.clone(),
+                        })
+                        .collect();
+
+                let analyzed = crate::table_builder::analyze_l2_to_l1_continuation_calls(
+                    &[root_call],
+                    &return_calls_for_builder,
+                    rollup_id,
+                );
+                let continuation = crate::table_builder::build_l2_to_l1_continuation_entries(
+                    &analyzed,
+                    cross_chain::RollupId::new(alloy_primitives::U256::from(rollup_id)),
+                    rlp_encoded_tx,
+                    cross_chain::TxOutcome::Success, // tx_reverts
+                );
+
+                tracing::info!(
+                    target: "based_rollup::proxy",
+                    call_idx = i,
+                    l1_entry_count = continuation.l1_entries.len(),
+                    return_call_count = my_return_calls.len(),
+                    "built combined simulation entries via table builder"
+                );
+
+                continuation.l1_entries
+            };
+
+            combined_entries.extend(entries);
+        }
+
+        if combined_entries.is_empty() {
+            tracing::warn!(
+                target: "based_rollup::proxy",
+                "combined entry building produced no L1 deferred entries"
+            );
+            break;
+        }
+
+        tracing::info!(
+            target: "based_rollup::proxy",
+            iteration,
+            total_entries = combined_entries.len(),
+            "built combined L1 deferred entries for all calls"
+        );
+
+        // Get L1 block context for proof signing.
+        let (block_number, block_hash, _parent_hash) =
+            match get_l1_block_context(client, l1_rpc_url).await {
+                Ok(ctx) => ctx,
+                Err(e) => {
+                    tracing::warn!(
+                        target: "based_rollup::proxy",
+                        %e,
+                        "failed to get L1 block context for combined simulation"
+                    );
+                    break;
+                }
+            };
+
+        let trace_block_number = block_number + 1;
+        let trace_parent_hash = block_hash;
+        let trace_block_timestamp = std::time::SystemTime::now()
+            .duration_since(std::time::UNIX_EPOCH)
+            .map(|d| d.as_secs())
+            .unwrap_or(0);
+
+        // Get verification key from Rollups contract.
+        let vk =
+            match get_verification_key(client, l1_rpc_url, rollups_address, rollup_id)
+                .await
+            {
+                Ok(v) => v,
+                Err(e) => {
+                    tracing::warn!(
+                        target: "based_rollup::proxy",
+                        %e,
+                        "failed to get verification key for combined simulation"
+                    );
+                    break;
+                }
+            };
+
+        // Sign ECDSA proof for combined postBatch.
+        let call_data_bytes = alloy_primitives::Bytes::new();
+        let entry_hashes = cross_chain::compute_entry_hashes(&combined_entries, vk);
+        let public_inputs_hash = cross_chain::compute_public_inputs_hash(
+            &entry_hashes,
+            &call_data_bytes,
+            trace_parent_hash,
+            trace_block_timestamp,
+        );
+
+        use alloy_signer::SignerSync;
+        let sig = match builder_key.sign_hash_sync(&public_inputs_hash) {
+            Ok(s) => s,
+            Err(e) => {
+                tracing::warn!(
+                    target: "based_rollup::proxy",
+                    %e,
+                    "failed to sign proof for combined L1 simulation"
+                );
+                break;
+            }
+        };
+        let sig_bytes = sig.as_bytes();
+        let mut proof_bytes = sig_bytes.to_vec();
+        if proof_bytes.len() == 65 && proof_bytes[64] < 27 {
+            proof_bytes[64] += 27;
+        }
+        let proof = alloy_primitives::Bytes::from(proof_bytes);
+
+        // Encode postBatch calldata.
+        let post_batch_calldata = cross_chain::encode_post_batch_calldata(
+            &combined_entries,
+            call_data_bytes,
+            proof,
+        );
+
+        // Build the traceCallMany bundle:
+        //   tx0: postBatch(combined_entries)
+        //   tx1: executeL2TX(rollupId, rlpTx)
+        // One executeL2TX handles all entries via scope resolution.
+        let builder_addr_hex = format!("{builder_address}");
+        let rollups_hex = format!("{rollups_address}");
+        let post_batch_data = format!("0x{}", hex::encode(post_batch_calldata.as_ref()));
+        let next_block = format!("{:#x}", trace_block_number);
+
+        // Encode executeL2TX calldata using typed ABI encoding (NEVER hardcode selectors).
+        let execute_l2tx_calldata = cross_chain::IRollups::executeL2TXCall {
+            rollupId: alloy_primitives::U256::from(rollup_id),
+            rlpEncodedTx: rlp_encoded_tx.to_vec().into(),
+        }
+        .abi_encode();
+        let execute_l2tx_data = format!("0x{}", hex::encode(&execute_l2tx_calldata));
+
+        let transactions = vec![
+            serde_json::json!({
+                "from": builder_addr_hex,
+                "to": rollups_hex,
+                "data": post_batch_data,
+                "gas": "0x1c9c380"
+            }),
+            serde_json::json!({
+                "from": builder_addr_hex,
+                "to": rollups_hex,
+                "data": execute_l2tx_data,
+                "gas": "0xc35000"
+            }),
+        ];
+
+        // All calls share a single executeL2TX trigger at index 1.
+        let call_trigger_tx_indices: Vec<usize> = (0..calls.len()).map(|_| 1).collect();
+
+        let expected_trace_count = transactions.len();
+
+        let trace_req = serde_json::json!({
+            "jsonrpc": "2.0",
+            "method": "debug_traceCallMany",
+            "params": [
+                [
+                    {
+                        "transactions": transactions,
+                        "blockOverride": {
+                            "number": next_block,
+                            "time": format!("{:#x}", trace_block_timestamp)
+                        }
+                    }
+                ],
+                null,
+                { "tracer": "callTracer" }
+            ],
+            "id": 5
+        });
+
+        let resp = match client.post(l1_rpc_url).json(&trace_req).send().await {
+            Ok(r) => match r.json::<Value>().await {
+                Ok(v) => v,
+                Err(e) => {
+                    tracing::warn!(
+                        target: "based_rollup::proxy",
+                        %e,
+                        "combined traceCallMany response parse failed"
+                    );
+                    break;
+                }
+            },
+            Err(e) => {
+                tracing::warn!(
+                    target: "based_rollup::proxy",
+                    %e,
+                    "combined traceCallMany request failed"
+                );
+                break;
+            }
+        };
+
+        // Extract traces from result.
+        let bundle_traces = match resp
+            .get("result")
+            .and_then(|r| r.get(0))
+            .and_then(|b| b.as_array())
+        {
+            Some(arr) if arr.len() >= expected_trace_count => arr,
+            _ => {
+                if let Some(error) = resp.get("error") {
+                    tracing::warn!(
+                        target: "based_rollup::proxy",
+                        ?error,
+                        "combined traceCallMany returned error"
+                    );
+                } else {
+                    let actual = resp
+                        .get("result")
+                        .and_then(|r| r.get(0))
+                        .and_then(|b| b.as_array())
+                        .map(|a| a.len())
+                        .unwrap_or(0);
+                    tracing::warn!(
+                        target: "based_rollup::proxy",
+                        expected = expected_trace_count,
+                        actual,
+                        "combined traceCallMany returned unexpected trace count"
+                    );
+                }
+                break;
+            }
+        };
+
+        // Check postBatch result (tx0).
+        let tx0_trace = &bundle_traces[0];
+        if tx0_trace.get("error").is_some() || tx0_trace.get("revertReason").is_some() {
+            let error_msg = tx0_trace
+                .get("error")
+                .and_then(|v| v.as_str())
+                .unwrap_or("unknown");
+            tracing::warn!(
+                target: "based_rollup::proxy",
+                error = error_msg,
+                iteration,
+                "postBatch reverted in combined L1 delivery simulation"
+            );
+            if iteration == 1 {
+                return None;
+            }
+            break;
+        }
+
+        // Parse each trigger trace to extract delivery output and return calls.
+        let mut new_return_calls_this_iteration: Vec<ReturnEdge> = Vec::new();
+
+        for (call_idx, &trigger_tx_idx) in call_trigger_tx_indices.iter().enumerate() {
+            let trigger_trace = &bundle_traces[trigger_tx_idx];
+
+            let (return_data, _delivery_failed) = extract_delivery_output_from_trigger_trace(
+                trigger_trace,
+                calls[call_idx].destination,
+            );
+
+            per_call_return_data[call_idx] = return_data.clone();
+            // Trigger simulation is unreliable for L2→L1 calls with placeholder
+            // state deltas. Always assume delivery succeeds — §4f + rewind
+            // handles real failures.
+            per_call_delivery_failed[call_idx] = false;
+
+            // Extract L1→L2 return calls from this trigger trace.
+            // Parent scope = this call's scope from L2 trace depth.
+            let call_scope: Vec<U256> = if calls.len() > 1 {
+                let mut s = if calls[call_idx].trace_depth <= 1 {
+                    vec![]
+                } else {
+                    vec![U256::ZERO; calls[call_idx].trace_depth]
+                };
+                s.push(U256::from(call_idx));
+                s
+            } else if calls[call_idx].trace_depth <= 1 {
+                vec![]
+            } else {
+                vec![U256::ZERO; calls[call_idx].trace_depth]
+            };
+            let new_returns = extract_l1_to_l2_return_calls(
+                client,
+                l1_rpc_url,
+                rollups_address,
+                trigger_trace,
+                rollup_id,
+                &call_scope,
+            )
+            .await;
+
+            // Tag return calls with parent_call_index so we know which trigger produced them.
+            for mut rc in new_returns {
+                rc.parent_call_index = cross_chain::ParentLink::Child(cross_chain::AbsoluteCallIndex::new(call_idx));
+                new_return_calls_this_iteration.push(rc);
+            }
+        }
+
+        // Filter out already-known return calls using count-based comparison.
+        // Supports legitimate duplicate return calls with identical
+        // (destination, data, value, source_address, parent_call_index) tuples.
+        // The CALL action hash includes value and sourceAddress, so two calls with
+        // different ETH values are distinct even if destination and data match.
+        let truly_new = filter_new_by_count(
+            new_return_calls_this_iteration,
+            &all_return_calls,
+            |a, b| {
+                a.destination == b.destination
+                    && a.data == b.data
+                    && a.value == b.value
+                    && a.source_address == b.source_address
+                    && a.parent_call_index == b.parent_call_index
+            },
+        );
+
+        // Check convergence: no new return calls AND return data stabilized (#254 item 7).
+        let return_data_changed = per_call_return_data != prev_per_call_return_data;
+        if truly_new.is_empty() && !return_data_changed {
+            tracing::info!(
+                target: "based_rollup::proxy",
+                iteration,
+                total_return_calls = all_return_calls.len(),
+                "combined L1 delivery simulation converged — no new return calls, data stable"
+            );
+            break;
+        }
+        if truly_new.is_empty() && return_data_changed {
+            tracing::info!(
+                target: "based_rollup::proxy",
+                iteration,
+                "combined L1 delivery: no new return calls but return data changed — re-iterating"
+            );
+            prev_per_call_return_data = per_call_return_data.clone();
+            continue;
+        }
+
+        tracing::info!(
+            target: "based_rollup::proxy",
+            iteration,
+            new_return_calls = truly_new.len(),
+            "discovered new return calls in combined delivery trace — re-simulating"
+        );
+
+        all_return_calls.extend(truly_new);
+
+        // Enrich return calls with L2 return data (#254 item 6).
+        super::l2_to_l1::enrich_return_calls_via_l2_trace(
+            client,
+            l2_rpc_url,
+            cross_chain_manager_address,
+            &mut all_return_calls,
+            rollup_id,
+        )
+        .await;
+
+        prev_per_call_return_data = per_call_return_data.clone();
+    }
+
+    // Build per-call results.
+    let mut results: Vec<(Vec<u8>, bool, Vec<ReturnEdge>)> = Vec::new();
+    for (i, _call) in calls.iter().enumerate() {
+        let my_return_calls: Vec<ReturnEdge> = all_return_calls
+            .iter()
+            .filter(|rc| rc.parent_call_index == cross_chain::ParentLink::Child(cross_chain::AbsoluteCallIndex::new(i)))
+            .cloned()
+            .collect();
+
+        results.push((
+            per_call_return_data[i].clone(),
+            per_call_delivery_failed[i],
+            my_return_calls,
+        ));
+    }
+
+    if !all_return_calls.is_empty() {
+        tracing::info!(
+            target: "based_rollup::proxy",
+            total_return_calls = all_return_calls.len(),
+            num_calls = calls.len(),
+            "combined L1 delivery simulation complete (return calls discovered)"
+        );
+    } else {
+        tracing::info!(
+            target: "based_rollup::proxy",
+            num_calls = calls.len(),
+            "combined L1 delivery simulation complete (no return calls)"
+        );
+    }
+
+    Some(results)
+}
+
+/// Simulate L2→L1 calls chained in a single `debug_traceCallMany` bundle.
+///
+/// Unlike `simulate_l1_delivery` which builds full [postBatch, createProxy, trigger]
+/// bundles, this function simulates DIRECT calls from the L1 proxy to the destination
+/// — lighter weight but sufficient for capturing per-call return data with state
+/// accumulation. Does NOT discover return calls (use `simulate_l1_combined_delivery`
+/// for that).
+///
+/// Returns `Vec<(return_data, call_success)>` with one entry per call.
+/// On any transport/parse failure, falls back to per-call independent simulation.
+#[allow(clippy::too_many_arguments)]
+pub(crate) async fn simulate_chained_delivery_l2_to_l1(
+    client: &reqwest::Client,
+    l1_rpc_url: &str,
+    _l2_rpc_url: &str,
+    rollups_address: Address,
+    rollup_id: u64,
+    _builder_address: Address,
+    _builder_private_key: &str,
+    calls: &[DiscoveredCall],
+    _trace_block_number: u64,
+    _trace_block_timestamp: u64,
+) -> Vec<(Vec<u8>, bool)> {
+    if calls.is_empty() {
+        return vec![];
+    }
+
+    // Step 1: Compute the L1 proxy address for the source.
+    // All identical calls share the same source_address, so compute once.
+    let source_address = calls[0].source_address;
+    let proxy_from = match compute_proxy_address_on_l1(
+        client,
+        l1_rpc_url,
+        rollups_address,
+        source_address,
+        rollup_id,
+    )
+    .await
+    {
+        Ok(addr) => addr,
+        Err(e) => {
+            tracing::warn!(
+                target: "based_rollup::proxy",
+                %e,
+                source = %source_address,
+                "chained L1 proxy address lookup failed — falling back to per-call simulation"
+            );
+            return fallback_per_call_l2_to_l1_simulation(client, l1_rpc_url, calls).await;
+        }
+    };
+
+    // Step 2: Check that all destinations have code on L1 (not EOAs).
+    for call in calls {
+        let code_req = serde_json::json!({
+            "jsonrpc": "2.0",
+            "method": "eth_getCode",
+            "params": [format!("{}", call.destination), "latest"],
+            "id": 1
+        });
+        let code_resp = match client.post(l1_rpc_url).json(&code_req).send().await {
+            Ok(r) => r,
+            Err(e) => {
+                tracing::warn!(
+                    target: "based_rollup::proxy",
+                    %e,
+                    "chained L2→L1: eth_getCode failed — falling back"
+                );
+                return fallback_per_call_l2_to_l1_simulation(client, l1_rpc_url, calls).await;
+            }
+        };
+        let code_body: Value = match code_resp.json().await {
+            Ok(b) => b,
+            Err(e) => {
+                tracing::warn!(
+                    target: "based_rollup::proxy",
+                    %e,
+                    "chained L2→L1: eth_getCode parse failed — falling back"
+                );
+                return fallback_per_call_l2_to_l1_simulation(client, l1_rpc_url, calls).await;
+            }
+        };
+        let code_hex = code_body
+            .get("result")
+            .and_then(|v| v.as_str())
+            .unwrap_or("0x");
+        if code_hex == "0x" || code_hex == "0x0" {
+            tracing::info!(
+                target: "based_rollup::proxy",
+                destination = %call.destination,
+                "chained L2→L1: destination is EOA — returning empty for all calls"
+            );
+            // If any destination is an EOA, return empty for ALL calls
+            // (all identical calls share the same destination).
+            return calls.iter().map(|_| (vec![], true)).collect();
+        }
+    }
+
+    // Step 3: Get L1 block context for blockOverride.
+    let (block_number, _block_hash, _parent_hash) =
+        match get_l1_block_context(client, l1_rpc_url).await {
+            Ok(ctx) => ctx,
+            Err(e) => {
+                tracing::warn!(
+                    target: "based_rollup::proxy",
+                    %e,
+                    "chained L2→L1: failed to get L1 block context — falling back"
+                );
+                return fallback_per_call_l2_to_l1_simulation(client, l1_rpc_url, calls).await;
+            }
+        };
+
+    let trace_block_number_val = block_number + 1;
+    let trace_block_timestamp_val = std::time::SystemTime::now()
+        .duration_since(std::time::UNIX_EPOCH)
+        .map(|d| d.as_secs())
+        .unwrap_or(0);
+
+    // Step 4: Build debug_traceCallMany request with ONE bundle containing N delivery
+    // calls. Each tx executes from the L1 proxy to the destination, seeing state
+    // effects from previous calls in the bundle.
+    let proxy_from_hex = format!("{proxy_from}");
+    let next_block = format!("{:#x}", trace_block_number_val);
+
+    let transactions: Vec<Value> = calls
+        .iter()
+        .map(|call| {
+            serde_json::json!({
+                "from": &proxy_from_hex,
+                "to": format!("{}", call.destination),
+                "data": format!("0x{}", hex::encode(&call.calldata)),
+                "value": format!("0x{:x}", call.value),
+                "gas": "0x2faf080"
+            })
+        })
+        .collect();
+
+    let trace_req = serde_json::json!({
+        "jsonrpc": "2.0",
+        "method": "debug_traceCallMany",
+        "params": [
+            [
+                {
+                    "transactions": transactions,
+                    "blockOverride": {
+                        "number": next_block,
+                        "time": format!("{:#x}", trace_block_timestamp_val)
+                    }
+                }
+            ],
+            null,
+            { "tracer": "callTracer" }
+        ],
+        "id": 99940
+    });
+
+    tracing::info!(
+        target: "based_rollup::proxy",
+        num_calls = calls.len(),
+        proxy = %proxy_from,
+        "chained L2→L1 delivery simulation: debug_traceCallMany with {} txs in one bundle",
+        calls.len()
+    );
+
+    // Step 5: Execute the simulation.
+    let resp = match client.post(l1_rpc_url).json(&trace_req).send().await {
+        Ok(r) => r,
+        Err(e) => {
+            tracing::warn!(
+                target: "based_rollup::proxy",
+                %e,
+                "chained L2→L1 simulation (debug_traceCallMany) request failed — falling back"
+            );
+            return fallback_per_call_l2_to_l1_simulation(client, l1_rpc_url, calls).await;
+        }
+    };
+    let body: Value = match resp.json().await {
+        Ok(b) => b,
+        Err(e) => {
+            tracing::warn!(
+                target: "based_rollup::proxy",
+                %e,
+                "chained L2→L1 simulation response parse failed — falling back"
+            );
+            return fallback_per_call_l2_to_l1_simulation(client, l1_rpc_url, calls).await;
+        }
+    };
+
+    // Step 6: Parse the response.
+    // Structure: result[0] is an array of N trace objects (one per tx in the bundle).
+    let traces = match body
+        .get("result")
+        .and_then(|r| r.get(0))
+        .and_then(|b| b.as_array())
+    {
+        Some(arr) if arr.len() == calls.len() => arr,
+        _ => {
+            let actual_len = body
+                .get("result")
+                .and_then(|r| r.get(0))
+                .and_then(|b| b.as_array())
+                .map(|a| a.len())
+                .unwrap_or(0);
+            tracing::warn!(
+                target: "based_rollup::proxy",
+                expected = calls.len(),
+                actual = actual_len,
+                "chained L2→L1 simulation returned unexpected trace count — falling back"
+            );
+            return fallback_per_call_l2_to_l1_simulation(client, l1_rpc_url, calls).await;
+        }
+    };
+
+    // Step 7: Extract per-call results from each trace.
+    let mut results = Vec::with_capacity(calls.len());
+    for (i, trace) in traces.iter().enumerate() {
+        let has_error = trace.get("error").is_some() || trace.get("revertReason").is_some();
+        let output_hex = trace.get("output").and_then(|v| v.as_str()).unwrap_or("0x");
+        let hex_clean = output_hex.strip_prefix("0x").unwrap_or(output_hex);
+        let output_bytes = hex::decode(hex_clean).unwrap_or_default();
+
+        let success = !has_error;
+
+        tracing::info!(
+            target: "based_rollup::proxy",
+            idx = i,
+            success,
+            return_data_len = output_bytes.len(),
+            "chained L2→L1 simulation: call {} result",
+            i
+        );
+
+        results.push((output_bytes, success));
+    }
+
+    results
+}
+
+/// Fallback: simulate each L2->L1 call independently via direct eth_call.
+///
+/// Used when the chained simulation fails. Runs each call as an independent
+/// `debug_traceCallMany` with a single tx — no state accumulation between calls.
+pub(crate) async fn fallback_per_call_l2_to_l1_simulation(
+    client: &reqwest::Client,
+    l1_rpc_url: &str,
+    calls: &[DiscoveredCall],
+) -> Vec<(Vec<u8>, bool)> {
+    tracing::info!(
+        target: "based_rollup::proxy",
+        num_calls = calls.len(),
+        "falling back to per-call L2→L1 simulation (no state accumulation)"
+    );
+    let mut results = Vec::with_capacity(calls.len());
+    for call in calls {
+        // Simple standalone trace for each call — from zero address to destination.
+        // This won't accumulate state but gives a safe fallback.
+        let trace_req = serde_json::json!({
+            "jsonrpc": "2.0",
+            "method": "debug_traceCallMany",
+            "params": [[{
+                "transactions": [{
+                    "from": format!("{}", Address::ZERO),
+                    "to": format!("{}", call.destination),
+                    "data": format!("0x{}", hex::encode(&call.calldata)),
+                    "value": format!("0x{:x}", call.value),
+                    "gas": "0x2faf080"
+                }]
+            }], null, { "tracer": "callTracer" }],
+            "id": 99941
+        });
+
+        let resp = match client.post(l1_rpc_url).json(&trace_req).send().await {
+            Ok(r) => match r.json::<Value>().await {
+                Ok(v) => v,
+                Err(_) => {
+                    results.push((vec![], true));
+                    continue;
+                }
+            },
+            Err(_) => {
+                results.push((vec![], true));
+                continue;
+            }
+        };
+
+        let trace = resp
+            .get("result")
+            .and_then(|r| r.get(0))
+            .and_then(|b| b.as_array())
+            .and_then(|arr| arr.first());
+
+        match trace {
+            Some(t) => {
+                let has_error = t.get("error").is_some() || t.get("revertReason").is_some();
+                let output_hex = t.get("output").and_then(|v| v.as_str()).unwrap_or("0x");
+                let hex_clean = output_hex.strip_prefix("0x").unwrap_or(output_hex);
+                let output_bytes = hex::decode(hex_clean).unwrap_or_default();
+                results.push((output_bytes, !has_error));
+            }
+            None => {
+                results.push((vec![], true));
+            }
+        }
+    }
+    results
 }
